Add APE errors handling and fix chip ID reporing error
* The error message will report the FEEID of the link and ChipID where the error was found.

Alpide Protocol Extension error will be reported via their APE name, if the error is fatal,
the cable data will be discared, otherwise just 0 padding will be skipped before trying to read
next chip data.
Since the APE error codes are set instead of the ChipHeader, the ChipID is not known at this stage,
so the error message will report only FEEID and ChipID = -1.
The unknown ChipID (-1) may be encountered also in case of general "unknown word" error.

+++ DataFormats/Detectors/ITSMFT/common/include/DataFormatsITSMFT/NoiseMap.h
@@ -54,7 +54,7 @@ class NoiseMap
   /// Get the noise level for this pixels
   float getNoiseLevel(int chip, int row, int col) const
   {
-    if (chip > mNoisyPixels.size()) {
+    if (chip > (int)mNoisyPixels.size()) {
       return 0;
     }
     const auto keyIt = mNoisyPixels[chip].find(getKey(row, col));
@@ -66,7 +66,7 @@ class NoiseMap
 
   void increaseNoiseCount(int chip, int row, int col)
   {
-    if (chip > mNoisyPixels.size()) {
+    if (chip > (int)mNoisyPixels.size()) {
       return;
     }
     mNoisyPixels[chip][getKey(row, col)]++;
@@ -117,16 +117,16 @@ class NoiseMap
 
   bool isNoisy(int chip, int row, int col) const
   {
-    return chip < mNoisyPixels.size() && (mNoisyPixels[chip].find(getKey(row, col)) != mNoisyPixels[chip].end());
+    return chip < (int)mNoisyPixels.size() && (mNoisyPixels[chip].find(getKey(row, col)) != mNoisyPixels[chip].end());
   }
 
-  bool isNoisy(int chip) const { return chip < mNoisyPixels.size() && !mNoisyPixels[chip].empty(); }
+  bool isNoisy(int chip) const { return chip < (int)mNoisyPixels.size() && !mNoisyPixels[chip].empty(); }
 
   // Methods required by the calibration framework
   void print();
   void fill(const gsl::span<const CompClusterExt> data);
   void merge(const NoiseMap* prev) {}
-  const std::map<int, int>* getChipMap(int chip) const { return chip < mNoisyPixels.size() ? &mNoisyPixels[chip] : nullptr; }
+  const std::map<int, int>* getChipMap(int chip) const { return chip < (int)mNoisyPixels.size() ? &mNoisyPixels[chip] : nullptr; }
 
  private:
   static constexpr int SHIFT = 10, MASK = (0x1 << SHIFT) - 1;
